// src/api/models/counsel-card-base-information-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BaseInfoDTO } from './base-info-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { CounselPurposeAndNoteDTO } from './counsel-purpose-and-note-dto';

/**
 * 
 * @export
 * @interface CounselCardBaseInformationRes
 */
export interface CounselCardBaseInformationRes {
    /**
     * 
     * @type {BaseInfoDTO}
     * @memberof CounselCardBaseInformationRes
     */
    'baseInfo'?: BaseInfoDTO;
    /**
     * 
     * @type {string}
     * @memberof CounselCardBaseInformationRes
     */
    'cardRecordStatus'?: CounselCardBaseInformationResCardRecordStatusEnum;
    /**
     * 
     * @type {CounselPurposeAndNoteDTO}
     * @memberof CounselCardBaseInformationRes
     */
    'counselPurposeAndNote'?: CounselPurposeAndNoteDTO;
}

export const CounselCardBaseInformationResCardRecordStatusEnum = {
    NotStarted: 'NOT_STARTED',
    InProgress: 'IN_PROGRESS',
    Completed: 'COMPLETED'
} as const;

export type CounselCardBaseInformationResCardRecordStatusEnum = typeof CounselCardBaseInformationResCardRecordStatusEnum[keyof typeof CounselCardBaseInformationResCardRecordStatusEnum];